requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
openai>=1.0.0
pandas>=1.3.3
tqdm>=4.62.3
ratelimit>=2.2.1
fake-useragent>=0.1.11
google-search-results>=2.4.1
SQLAlchemy>=1.4.23